refactor($dto): reorganize package layout for dto
