+++ src/main/java/racingcar/dto/request/TryCount.java
@@ -1,4 +1,4 @@
-package racingcar.dto;
+package racingcar.dto.request;
 
 import java.util.Objects;
